local_backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-socketio==5.16.1
pydantic==2.10.5
motor==3.6.0
